+++ howhowSpecialProject/src/main/java/com/howhow/shopping/controller/ShoppingCartController.java
@@ -63,7 +63,7 @@ public class ShoppingCartController {
 			int courseID = shoppingCart.getCourseBasic().getCourseID();
 			CourseBasic course = cService.findByID(courseID);
 			String courseName = course.getCourseName();
-			int price = course.getPrice();
+			long price = course.getPrice();
 			double discount = course.getDiscount();
 			int discountPrice = (int) (price * discount);
 			double rank = rankUtilsByCourseID(courseID);
